+++ Game/Hidden_Armada/Hidden_Armada/source/game/objects/PlayerShip.cpp
@@ -55,25 +55,25 @@ void PlayerShip::Update( float _dt )
 
 	if(GetPos().x - GetSize() < 0)
 	{
-		SetPos(D3DXVECTOR2(0 + GetSize(), GetPos().y));
+		SetPos(D3DXVECTOR2((float)(0 + GetSize()), GetPos().y));
 		//SetVel(D3DXVECTOR2(-GetVel().x, GetVel().y));
 	}
 	
 	if(GetPos().x + GetSize() * 4 > 1024 * 4)
 	{
-		SetPos(D3DXVECTOR2(1024*4 - GetSize() * 4, GetPos().y));
+		SetPos(D3DXVECTOR2((float)(1024*4 - GetSize() * 4), GetPos().y));
 		//SetVel(D3DXVECTOR2(-GetVel().x, GetVel().y));
 	}
 
 	if(GetPos().y - GetSize() < 0)
 	{
-		SetPos(D3DXVECTOR2(GetPos().x, 0 + GetSize()));
+		SetPos(D3DXVECTOR2(GetPos().x, (float)(0 + GetSize())));
 		//SetVel(D3DXVECTOR2(GetVel().x, -GetVel().y));
 	}
 	
 	if(GetPos().y + GetSize() * 4 > 768 * 4)
 	{
-		SetPos(D3DXVECTOR2(GetPos().x, 768*4 - GetSize() * 4));
+		SetPos(D3DXVECTOR2(GetPos().x, (float)(768*4 - GetSize() * 4)));
 		//SetVel(D3DXVECTOR2(GetVel().x, -GetVel().y));
 	}
